requirements.txt
asgiref==3.5.2
astroid==2.9.3
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.4
distlib==0.3.6
Django==3.2
django-filter==22.1
djangorestframework==3.14.0
filelock==3.8.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.4.1
mccabe==0.6.1
numpy==1.23.3
pandas==1.5.0
platformdirs==2.4.1
pylint==2.12.2
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.4
style==1.1.0
toml==0.10.2
tzdata==2022.2
update==0.0.1
urllib3==1.26.12
virtualenv==20.16.5
wrapt==1.13.3
